refactor(contracts): extract CreatedListingResponseDto
