+++ apps/mobile/src/features/listings/infrastructure/mappers/CreatedListingMapper.ts
@@ -1,14 +1,5 @@
+import { CreatedListingResponseDto } from '@esparex/contracts';
 import { CreatedListing } from '../../domain/CreatedListing';
-
-/**
- * CreatedListingResponseDto — strongly typed contract for backend creation responses.
- */
-export interface CreatedListingResponseDto {
-  id: string | number;
-  slug?: string;
-  status?: string;
-  moderationStatus?: string;
-}
 
 /**
  * CreatedListingMapper — pure mapper transforming raw API DTO responses into CreatedListing domain value objects.
