+++ packages/contracts/src/v1/listings/dto/CreatedListingResponseDto.ts
@@ -1,0 +1,9 @@
+/**
+ * CreatedListingResponseDto — strongly typed contract for backend creation responses.
+ */
+export interface CreatedListingResponseDto {
+  id: string | number;
+  slug?: string;
+  status?: string;
+  moderationStatus?: string;
+}
