+++ packages/contracts/src/v1/listings/dto/index.ts
@@ -3,4 +3,5 @@ export * from './service';
 export * from "./ListingQueryParams";
 export * from './PostAdDraft';
 export * from './CreateListingRequest';
+export * from './CreatedListingResponseDto';
